Импорт модулей в main.js
- импорт модулей в точку входа main.js
- подключение main.js в index.html.
