+++ js/editor-file.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line eol-last
+//Редактирование изображения и ограничения, накладываемые на поля
