+++ js/filter.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line eol-last
+//Фильтрация изображений от других пользователей.
